# backend/openapi.yaml
openapi: 3.0.3
info:
  title: EliteDial API
  description: Backend API for the EliteDial predictive dialer and call center platform.
  version: 1.0.0
servers:
  - url: /api
    description: Main API (all routes except webhooks)
  - url: /
    description: Root (health check and webhook endpoints)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [agent, supervisor, admin]
        status:
          type: string
          enum: [available, break, offline, on-call]
        extension:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
...
    PhoneNumber:
      type: object
      properties:
        id:
          type: string
          format: uuid
        number:
          type: string
        label:
          type: string
          nullable: true
        type:
          type: string
          enum: [local, toll-free, mobile]
        assignedTo:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    DNCEntry:
      type: object
      properties:
        id:
          type: string
          format: uuid
        phoneNumber:
          type: string
        reason:
          type: string
          nullable: true
        addedBy:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    QueueConfig:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        holdTimeout:
          type: integer
        overflowAction:
          type: string
          enum: [voicemail, virtual-agent, continue-hold]
        holdMusicUrl:
          type: string
          nullable: true
        maxQueueSize:
          type: integer
        isActive:
          type: boolean
    DispositionCode:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        label:
          type: string
        category:
          type: string
          enum: [payment, promise, skip, general]
        isActive:
          type: boolean
    APIKey:
      type: object
      properties:
        id:
          type: string
          format: uuid
        key:
          type: string
        label:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        lastUsed:
          type: string
          format: date-time
          nullable: true
    WebhookConfig:
      type: object
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
        secret:
          type: string
        events:
          type: string
          description: Comma-separated event names
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
security:
  - bearerAuth: []
paths:
...
  /admin/agents/{id}:
    put:
      tags: [Admin]
      summary: Update an agent profile (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                email:
                  type: string
                  format: email
                role:
                  type: string
                  enum: [agent, supervisor, admin]
                extension:
                  type: string
      responses:
        '200':
          description: Agent updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    delete:
      tags: [Admin]
      summary: Delete an agent (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Agent deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
  /admin/agents/{id}/reset-password:
    post:
      tags: [Admin]
      summary: Reset an agent's password (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [newPassword]
              properties:
                newPassword:
                  type: string
      responses:
        '200':
          description: Password reset
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
  /admin/phones:
    get:
      tags: [Admin]
      summary: List all phone numbers (admin)
      responses:
        '200':
          description: Phone number list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PhoneNumber'
    post:
      tags: [Admin]
      summary: Add a phone number (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [number]
              properties:
                number:
                  type: string
                label:
                  type: string
                type:
                  type: string
                  enum: [local, toll-free, mobile]
                assignedTo:
                  type: string
      responses:
        '201':
          description: Phone number created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhoneNumber'
  /admin/phones/{id}:
    delete:
      tags: [Admin]
      summary: Delete a phone number (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Phone deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
  /admin/dnc:
    get:
      tags: [Admin]
      summary: List DNC entries (admin, paginated)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: DNC list
          content:
            application/json:
              schema:
                type: object
                properties:
                  entries:
                    type: array
                    items:
                      $ref: '#/components/schemas/DNCEntry'
                  total:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer
    post:
      tags: [Admin]
      summary: Add a phone number to the DNC list (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phoneNumber]
              properties:
                phoneNumber:
                  type: string
                reason:
                  type: string
      responses:
        '201':
          description: DNC entry added
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
  /admin/dnc/{phoneNumber}:
    delete:
      tags: [Admin]
      summary: Remove a phone number from the DNC list (admin)
      parameters:
        - name: phoneNumber
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: DNC entry removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
  /admin/dnc/check/{phoneNumber}:
    get:
      tags: [Admin]
      summary: Check if a phone number is on the DNC list
      parameters:
        - name: phoneNumber
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: DNC check result
          content:
            application/json:
              schema:
                type: object
                properties:
                  phoneNumber:
                    type: string
                  isDNC:
                    type: boolean
  /admin/dnc/import:
    post:
      tags: [Admin]
      summary: Bulk-import phone numbers to the DNC list (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [numbers]
              properties:
                numbers:
                  type: array
                  items:
                    type: string
                reason:
                  type: string
      responses:
        '200':
          description: Import count
          content:
            application/json:
              schema:
                type: object
                properties:
                  imported:
                    type: integer
...
  /admin/queues/{id}:
    put:
      tags: [Admin]
      summary: Update a queue configuration (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                holdTimeout:
                  type: integer
                overflowAction:
                  type: string
                  enum: [voicemail, virtual-agent, continue-hold]
                holdMusicUrl:
                  type: string
                maxQueueSize:
                  type: integer
                isActive:
                  type: boolean
      responses:
        '200':
          description: Queue updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueueConfig'
  /admin/dispositions:
    get:
      tags: [Admin]
      summary: List active disposition codes
      responses:
        '200':
          description: Disposition code list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DispositionCode'
    post:
      tags: [Admin]
      summary: Create a disposition code (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, label]
              properties:
                code:
                  type: string
                label:
                  type: string
                category:
                  type: string
                  enum: [payment, promise, skip, general]
      responses:
        '201':
          description: Disposition created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DispositionCode'
  /admin/api-keys:
    get:
      tags: [Admin]
      summary: List API keys (admin)
      responses:
        '200':
          description: API key list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/APIKey'
    post:
      tags: [Admin]
      summary: Create an API key (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                label:
                  type: string
      responses:
        '201':
          description: API key created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIKey'
  /admin/webhooks:
    get:
      tags: [Admin]
      summary: List webhook configurations (admin)
      responses:
        '200':
          description: Webhook list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WebhookConfig'
    post:
      tags: [Admin]
      summary: Create a webhook configuration (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url, secret, events]
              properties:
                url:
                  type: string
                  format: uri
                secret:
                  type: string
                events:
                  type: string
                  description: Comma-separated event names (e.g. call.answered,call.completed)
      responses:
        '201':
          description: Webhook created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookConfig'

  # ───────────────────────────────────────────────
  # System
  # ───────────────────────────────────────────────
...